requirements.txt
torch
ultralytics
transformers
Pillow
clip
matplotlib
numpy